Move Results and QueryInterrupted to datasette/database.py
